Split functions into their own files.
Runs just like the original code.
