+++ correlations.dat
@@ -1,15 +1,15 @@
-1.272198
-0.986792
-0.766599
-0.595660
-0.462045
-0.361380
-0.285460
-0.215798
-0.174051
-0.140573
-0.114770
-0.098883
-0.078242
-0.057750
+1.292425
+1.008481
+0.777081
+0.607362
+0.473575
+0.373384
+0.291646
+0.224219
+0.171028
+0.129042
+0.104830
+0.091779
+0.068933
+0.058559
 0.000000
