--- mhmath.h
@@ -1,13 +1,0 @@
-/* mhmath.h
- * Header file containing commonly used math routines
- * Brad Theilman September 2015
- */
- 
- #ifndef MHMATH_H
- #define MHMATH_H
- 
- double rand_in_range(double min, double max);
- double* generate_random_initial_path(int nt, double min, double max);
- 
- 
- #endif
